pyproject.toml
[project]
name = "rag-q-and-a-conversation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]
